feat: add @gjsify/adwaita-core headless seed
Implement ADR 0004 step 1: create the headless Adwaita behavior
package and move the first two platform-neutral pieces out of
@gjsify/adwaita-nativescript:

- breakpoint condition grammar/parser/evaluator + the
  transition-only AdwBreakpoint apply/unapply state machine
- the light/dark color-scheme observable + the default icon-fg
  constants

The NS package keeps only its platform binding (addBreakpoints view
wiring) and re-exports the moved surface from @gjsify/adwaita-core,
so no consumer-visible break (easy6502 app-android imports keep
working unchanged). Specs ported to the core package and run on GJS
+ Node (53 each); NS re-export smoke specs pin the compat chain
(121 each on GJS + Node).

Placement: packages/web/adwaita-core/ next to the design-identity
siblings (adwaita-web/-fonts/-icons). Runtimes all-polyfill,
tier 2, no /register, pure named exports.

RELEASE BLOCKER: new npm package — manual first publish + gjsify
trust required before the next release (see STATUS.md Open TODOs).
