+++ packages/web/adwaita-core/src/test.mts
@@ -1,0 +1,6 @@
+import { run } from '@gjsify/unit';
+
+import breakpointTestSuite from './breakpoint.spec.js';
+import colorSchemeTestSuite from './color-scheme.spec.js';
+
+run({ breakpointTestSuite, colorSchemeTestSuite });
